requirements.txt
flask==3.0.3
waitress==3.0.0
pillow==10.4.0